admin_panel/requirements.txt
flask==3.0.0
flask-cors==4.0.0
motor==3.3.2
pymongo==4.5.0
python-dotenv==1.0.0
gunicorn==21.2.0 